% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_template.R
\name{make_qtw_template}
\alias{make_qtw_template}
\title{Make Quarto Website template}
\usage{
make_qtw_template(path, template_option)
}
\arguments{
\item{path}{Path of the R project}

\item{template_option}{Template option to choose.}
}
\value{
Returns nothing.
}
\description{
This function makes an R project that includes 3 basic templates
that are commonly used by academics.
}
\examples{
\dontrun{
  make_qtw_template(path = tempdir(), template_option = "Minimal")
}
}
